refactor: reorg dir
